requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==2.4.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
dnspython==2.8.0
duckduckgo_search==8.1.1
executing==2.1.0
fastapi==0.128.0
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.54.1
frozenlist==1.8.0
fsspec==2024.12.0
future==1.0.0
google-auth==2.47.0
google-genai==1.57.0
graphviz==0.20.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.1
idna==3.11
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==1.2.3
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.1.3
langchain-mongodb==0.10.0
langchain-text-splitters==1.1.0
langchain-voyageai==0.3.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.2
langsmith==0.6.2
lark==1.3.1
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
primp==0.15.0
prompt_toolkit==3.0.48
propcache==0.4.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.18.0
pymongo==4.16.0
pymongo-search-utils==0.1.0
pyparsing==3.2.0
pypdf==6.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.1
PyYAML==6.0.3
pyzmq==26.2.0
reportlab==4.4.7
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
sympy==1.13.1
tenacity==9.1.2
tokenizers==0.22.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
voyageai==0.3.7
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
